+++ packages/docs/.storybook/main.ts
@@ -1,30 +1,30 @@
-import { join, dirname } from "path";
-import { StorybookConfig } from "@storybook/react-vite";
-import remarkGfm from "remark-gfm";
-import svgrPlugin from "vite-plugin-svgr";
-
+import { join, dirname } from 'path'
+import { StorybookConfig } from '@storybook/react-vite'
+import remarkGfm from 'remark-gfm'
+import svgrPlugin from 'vite-plugin-svgr'
+import tsConfigPaths from 'vite-tsconfig-paths'
 /**
  * This function is used to resolve the absolute path of a package.
  * It is needed in projects that use Yarn PnP or are set up within a monorepo.
  */
 function getAbsolutePath(value: string) {
-  return dirname(require.resolve(join(value, "package.json")));
+  return dirname(require.resolve(join(value, 'package.json')))
 }
 
 const config: StorybookConfig = {
   stories: [
-    "../src/pages/**/*.mdx",
-    "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)",
+    '../src/pages/**/*.mdx',
+    '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)',
   ],
   addons: [
-    getAbsolutePath("@storybook/addon-links"),
-    getAbsolutePath("@storybook/addon-essentials"),
-    getAbsolutePath("@storybook/addon-onboarding"),
-    getAbsolutePath("@storybook/addon-interactions"),
-    getAbsolutePath("@storybook/addon-backgrounds"),
-    getAbsolutePath("storybook-dark-mode"),
+    getAbsolutePath('@storybook/addon-links'),
+    getAbsolutePath('@storybook/addon-essentials'),
+    getAbsolutePath('@storybook/addon-onboarding'),
+    getAbsolutePath('@storybook/addon-interactions'),
+    getAbsolutePath('@storybook/addon-backgrounds'),
+    getAbsolutePath('storybook-dark-mode'),
     {
-      name: getAbsolutePath("@storybook/addon-docs"),
+      name: getAbsolutePath('@storybook/addon-docs'),
       options: {
         mdxPluginOptions: {
           mdxCompileOptions: {
@@ -34,23 +34,25 @@ const config: StorybookConfig = {
       },
     },
   ],
-  staticDirs: ["../public"],
+  staticDirs: ['../public'],
   framework: {
-    name: getAbsolutePath("@storybook/react-vite") as "@storybook/react-vite",
+    name: getAbsolutePath('@storybook/react-vite') as '@storybook/react-vite',
     options: {},
   },
   docs: {
-    autodocs: "tag",
+    autodocs: 'tag',
   },
   viteFinal: (config, { configType }) => {
     config.plugins = [
       ...(config.plugins || []),
       svgrPlugin({ svgrOptions: { icon: true } }),
-    ];
-    if (configType === "PRODUCTION") {
-      config.base = "/click-tools-ui/";
+
+      tsConfigPaths(),
+    ]
+    if (configType === 'PRODUCTION') {
+      config.base = '/click-tools-ui/'
     }
-    return config;
+    return config
   },
-};
-export default config;
+}
+export default config