+++ packages/react/src/components/Input/index.tsx
@@ -3,20 +3,20 @@ import { tv, VariantProps } from 'tailwind-variants'
 
 const inputStyle = tv({
   base: [
-    'ct-flex ct-w-full ct-rounded-md ct-border ct-px-3 ct-py-2',
-    'ct-text-md ct-text-primary-900',
-    'ct-bg-transparent ct-shadow-sm ct-transition-colors',
-    'file:ct-border-0 file:ct-bg-transparent file:ct-text-sm file:ct-font-medium',
-    'placeholder:ct-text-neutral-500',
-    'focus:ct-transition-shadow focus-visible:ct-outline-none',
-    'focus-visible:ct-ring-1',
-    'disabled:ct-cursor-not-allowed',
-    'read-only:ct-cursor-default',
+    'flex w-full rounded-md border px-3 py-2',
+    'text-md text-primary-900',
+    'bg-transparent shadow-sm transition-colors',
+    'file:border-0 file:bg-transparent file:text-sm file:font-medium',
+    'placeholder:text-neutral-500',
+    'focus:transition-shadow focus-visible:outline-none',
+    'focus-visible:ring-1',
+    'disabled:cursor-not-allowed',
+    'read-only:cursor-default',
   ],
   variants: {
     error: {
-      true: 'ct-border-error-500 focus-visible:ct-ring-error-500',
-      false: 'ct-border-primary-900 focus-visible:ct-ring-primary-900',
+      true: 'border-error-500 focus-visible:ring-error-500',
+      false: 'border-primary-900 focus-visible:ring-primary-900',
     },
   },
   defaultVariants: {